Move clang/clang++ specification to CMakeLists
This ensures that the native runtime will be built with clang
regardless of the default c/c++ compiler.

Currently, building the nativeruntime does not work with GCC due to
libstddc++ linker errors. Building with clang is sufficient for now as
it's available on all platforms. At some point we can revisit why the
GCC linking phase doesn't work.