Fix UnsupportedOperationException if a leaked Activity.recreate is called
Previously, ActivityController.recreate determined the current state of the
underlying Activity by calling the global

  ActivityLifecycleMonitorRegistry.getLifecycleStageOf

This method is unreliable for leaked Activities as a new
ActivityLifecycleMonitor is created before each test. This means that, for a
leaked Activity, the ActivityLifecycleMonitor from the previous test has
knowledge if its lifecycle state, not the new one.  When the Activity state was
queried using the current ActivityLifecycleMonitor, an
UnsupportedOperationException occurred.

To fix this, avoid using ActivityLifecycleMonitor as the source of truth for
Activity State in ActivityController, and instead add a new member variable to
ActivityController that maintains the state. Having ActivityController rely on
ActivityLifecycleMonitor for Activity state is circuitous -- AndroidX Test is
already using ActivityController to drive Activity lifecycles, so the
source-of-truth state should exist inside of ActivityController itself.

PiperOrigin-RevId: 410168815