Use Activity's member Instrumentation in ActivityController
Previously, ActivityController was using the global Instrumentation object,
which is replaced after each test. This contributed to an issue where an
Activity that leaked across tests caused illegal states to exist in
ActivityLifecycleMonitor. Instead, use the Activity's own Instrumentation
object, which will have a consistent view of the Activity's lifecycle.

PiperOrigin-RevId: 410271639