Add ActivityController.close that transitions Activity to destroyed state
Update ActivityController to implement AutoCloseable, and add a 'close'
method that is lifecycle-aware and transitions the underlying Activity
to the destroyed state, freeing all resources and making the Activity
eligible for gc.

This is a convenient way to ensure that ActivityControllers can be freed
without having to manage the underlying Activity lifecycles.

It also enables ActivityController to be managed using
try-with-resources.

PiperOrigin-RevId: 410291892