Only using AttributionSource for compile SDK 31 and above
We have tried to use Object to replace AttributionSource as type of
second input parameter for noteProxyOpNoThrow(int, AttributeSource,
String, bool) to make sure ShadowAppOpsManager can be loaded and used
when compile SDK is less than 31. But changed method's siganture doesn't
match origin method signature, and it will be used as shadow method. To
fix this problem, this CL uses looseSignatures for ShadowAppOpsManager,
and chagnes all input parameters' type to Object to meet
looseSignatures' requirement.

Signed-off-by: utzcoz <utzcoz@outlook.com>