Add a task to fetch nativeruntime artifacts from a GitHub CI run
Add support for a new system property, NATIVERUNTIME_ACTION_RUN_ID, that
can be used to fetch the native runtime artifacts from the artifacts of
a GitHub action run. This makes it easier and more automated to add the
set of nativeruntime shared libraries to build an uber JAR.

Fetching artifacts using an API requires a GitHub username and account
token. These can be specified from the command line using environment
variables.

Long-term, this logic could be relocated to a GitHub workflow, if a
workflow exists to perform releases.