Switch to Ninja for building the native runtime
The version of `make` provided in the MSYS2 does not work well with CMake on
Windows. This makes it impossible to do incremental nativeruntme builds on
Windows.

See https://stackoverflow.com/questions/2401976 for more details on the issue.

Switch to using Ninja everywhere, which seems to be more reliable across
platforms.

PiperOrigin-RevId: 413717789