Use real Android code for ContentProviderClient.release
This fixes two issues:
1) Because 'release' was shadowed, the underlying CloseGuard for the
ContentProviderClient was never closed, leading to CloseGuard errors that could
not be suppressed.

2) The behavior of 'release' was not correct for SDK > 23, where calling it
multiple times does not result in an IllegalStateException.

Also, update ShadowContentProviderClientTest to use a real ContentProvider, and
remove the interaction test. Trying to release a ContentProviderClient for a
mock ContentProvider results in an exception.

PiperOrigin-RevId: 413795247