Use invokespecial when invoking shadow constructors
Previously, invokevirtual was used to invoke shadow __constructor__ methods.
This caused problems for shadowing constructor hierarchies with the same
signature.

Consider this contrived example:

  public class A {
    public A() {
      System.out.println("A!");
    }
  }

  @Instrument
  public class B extends A {
    public B() {
      super();
      System.out.println("B!");
    }
  }

And the corresponding shadows:

  @Implements(A.class)
  public class ShadowA {
    @RealObject protected A realA;

    @Implementation
    protected void __constructor__() {
      invokeConstructor(A.class, realA);
    }
  }

  @Implements(B.class)
  public class ShadowB extends ShadowA {
    @RealObject protected B realB;

    @Implementation
    protected void __constructor__() {
      invokeConstructor(B.class, realB);
    }
  }

Previously, when running `new B()`, it would print "B!" twice. This is
because when Robolectric's ShadowWrangler called the __constructor__ methods,
it would use virtual method dispatch, which resulted in B's __constructor__
method being called twice.

Update ShadowWrangler to use INVOKESPECIAL when invoke shadow __constructor__
methods. This avoids a virtual invocation for the method, which ensures that
the correct constructor method is called when constructor hierarchies are
shadowed.

PiperOrigin-RevId: 420507245