Fix touch event dispatch to not go to out of bounds non touch modal windows
If a touchable window is not touch modal (which includes window that are not focusable) then touch events outside of its bounds should be allowed to go to windows behind it.

This addresses a bug when dispatching touch events after focusing on an EditText which uses a PopupWindow for the cursor (presumably to catch touch events on the cursor), but the old dispatch behavior would send all touch events to the PopupWindow even though it was not touch modal.

PiperOrigin-RevId: 424525340