Return the original method's test name
The `FrameworkMethod` is recreated when running the method block statements so that it references the method class loaded by the Robolectric classloader. Plumb through the `testName` on the runner helper to return the name of the original framework method, this will potentially include the sdk parameter suffix if the test is running multiple sdks (e.g. test_name[30])

PiperOrigin-RevId: 428652762