Add shadows for ParcelFileDescriptor detachFd and adoptFd.
ParcelFileDescriptor detachFd and adoptFd method are about the ownership of a raw native fd in Android.

In shadow implementation, it's almost impossible to handle the native fd.
So the ownership transfering is simplified by "OpenedFileTable" managed and used only for PFD. It tracks all the
opened file that was opened by PFD's open method.

The "OpenedFileItem" is created when a file is opened by PFD. And it is removed when a PFD object is closed.
The existed "OpenedFileItem" is cached in memory when a PFD has been detached, and wait for the next adopt as a new PFD.
So when the new PFD can be created and restored from the existed "OpenedFileItem" which has a corresponding raw fd mapping.

Note that the whole PFD implementations are based on an internal attribute called `fd` of FileDescriptor class, but the `fd`
 is only available in *nix system. In Windows system, the `handle` attribute should be used.

See javadoc of FileDescriptor at:
https://github.com/openjdk/jdk/blob/master/src/java.base/share/classes/java/io/FileDescriptor.java

It means that some shadow methods related to file descriptor can not work in Windows, like getFd, detachFd, adoptFd.

Thus, the corresponding tests should be failed explictly in Windows platform.

Signed-off-by: ZSmallX <diosmallx@gmail.com>