Fix bitmap mutable, premutliplied, and alpha properties.
Defaults mutable, premultiplied to true and fix return values of alpha and premultiplied based on the config.

Real bitmaps in Android are typically mutable by default unless created using the colors int array factory method (which is used internally by the `BitmapDecoder` factory methods, update this to be true.

Fix the alpha property, which defaults to true in the constructor variants that do not explicitly specify it, and was previously being ignored in the overloaded constructors in the shadow.

Real bitmaps in Android are typically defaulted to premultiplied, non-premultiplied bitmaps cannot be drawn to the canvas and throw if tried. Android bitmaps actually retain the flag a "requested" because it can be false based on the config of the bitmap even if it is requested. The default behavior of the `ImageDecoder`, which creates bitmaps is to default premultiplied to true, and it's rarely overridden.

This is a pre-requisite to introducing some more realistic behavior to the canvas (which expects premultipled, mutable bitmaps), which in turn is a pre-requisite to defaulting activity windows to "visible".

PiperOrigin-RevId: 432600959