Don't decode image bytes as utf-8 encoded string
This is non-sensical, image bytes are not utf-8 encoded string bytes, and it's causing stack size issues for large images. Instead just put the length of the byte array in the description. Move the utf-8 decoding into the accessor for the deprecated `provideWidthAndHeightHints` methods. At some point this can be removed.

PiperOrigin-RevId: 434050940