Make ClassValueMap.get a one-liner and make it thread safe
We have been seeing NPEs when accessing the value from a ClassValueMap, e.g.

    java.lang.NullPointerException
      at org.robolectric.internal.bytecode.ShadowWrangler.injectReflectorObjectOn(ShadowWrangler.java:520)
      at android.graphics.Paint.$$robo$init(Paint.java)
      at android.graphics.Paint.<init>(Paint.java)

Replace ClassValueMap.get with a simple call to map.computeIfAbsent. Also
update the map to use a synchronized map. One possibility is that there
are race conditions that may have existed between calls to 'containsKey' and
'get'. With a combination of a synchronizedMap and computeIfAbsent, these races
are not possible any more. Add a concurrency stress test as well.

PiperOrigin-RevId: 436653896