Add READ_PHONE_STATE permission check to ShadowTelephonyManager
The real implementation of TelephonyManager requires the READ_PHONE_STATE permission when calling some methods (such as TelephonyManager#getServiceState and TelephonyManager#getImei). This change adds a check for that permission in the corresponding shadow's method so that tests relying on ShadowTelephonyManager can be written to take into consideration whether the permission has been granted or not.

PiperOrigin-RevId: 437051360