Indicate windows are visible by default (flag enabled)
Should more closely match real world behavior, the previous behavior was avoid executing a number of code paths that would run for visible windows. For now this is not enabled and is added based on a system property to allow for test migration. This flag is temporary and will be default enabled in a subsequent change, and finally removed entirely.

PiperOrigin-RevId: 437058543