Add shadow implementation of AccessibilityService.takeScreenshot for R+
Adds an implementation that immediately completes the callback with an
empty hardware buffer. Also add methods to the shadow to set and unset
error conditions, allowing the callback to instead complete with the
error code, to test code paths that respond to that.

Before this implementation existed, invoking takeScreenshot did nothing.
The callback would never be invoked, meaning it was impossible to test
any codepaths that are dependent on the callback being invoked.

PiperOrigin-RevId: 440205427