Add methods to `ShadowChoreographer` to configure vsync behavior
Previously the frame delay in `ShadowDisplayEventReceiver` was hardcoded to 1ms in unpaused mode, and configurable via `ShadowChoreographer.setPostFrameCallbackDelay` in paused mode. Split apart this method into `ShadowChoreographer.setFrameDelay` and `ShadowChoreographer.setPaused` to allow configuring the frame delay in both paused and unpaused modes and to allow toggling the paused mode without specifying a frame delay.

Uses `setFrameDelay` terminology from the `Choreographer` but uses `Duration` to remove unit ambiguity (the `ShadowChoreographer.setFrameInterval` actually takes a nanosecond time). Uses the "paused" terminology that the looper uses to configure the paused mode of the display events.

PiperOrigin-RevId: 445259956