Add looper mode assertions to `ShadowChoreographer` methods
It's error prone to call these methods with the wrong looper mode as they don't do anything, assert that the looper mode is supported. The caller can easily wrap the method call in a looper mode check if a test must support multiple looper modes (which is unusual).

Fixes a bug in the `AnimationRunner` where it was not incrementing the elapsed time for the animation by the correct increment for the paused looper.

PiperOrigin-RevId: 445298680