Improve the fidelity of Canvas.getClipBounds
Previously, it would always return the zero-rect.

Update it to at least return the bounds of the bitmap
that backs the Canvas.

Note that this does not support the various Canvas.clip* functions,
but it is an improvement over the previous version.

It may be possible to better support Canvas.clip* using AWT's Path2D.

PiperOrigin-RevId: 445435525