Fix dispatch order for windows in `LocalUiController`
The sort order was backwards for the index into the view roots list, and this updated behavior matches the behavior of Espresso's `RootMatcher`, previously Espresso would match a different root to the one Robolectric dispatched events to, which caused for confusing behavior. This is demonstrated by opening two dialogs, the events should go to the most recently opened dialog. Adds a test to verify the behavior is consistent with Android.

Filter the roots to only include roots for started activities, this also more closely matches Espresso's default root picking behavior, which picks resumed activities. The important thing is to filter out stopped activities which are not visible by definition, so should not receive events. I would suspect actually in real Android these view roots might already be removed from the window manager when the activity is stopped, but currently in Robolectric they are not (and I suspect the note about the TODO probably would follow the same logic, a stopped and removed window would have its child windows automatically removed too).

PiperOrigin-RevId: 446351635