Add shadow `UiAutomation` and implement `setAnimationScale`
This is a new API in Android T that is also useful to have backwards compatible support through the shadow. This will also make it slightly easier to add support for other shadow methods on `UiAutomation` in the future.

https://developer.android.google.cn/reference/android/app/UiAutomation#setAnimationScale(float)

Remove the public access from `ShadowValueAnimator.setDurationScale` and instead propagate the value from `Settings.Global.ANIMATOR_DURATION_SCALE`, this makes these objects a little more consistent, and the `UiAutomation` shadow provides access to set this that in T will be instrumentation compatible. This setup is similar to how real Android works where the value is pushed to `ValueAnimator`

PiperOrigin-RevId: 449497046