Use a MultiMap when loading ShadowProviders in ShadowMap
Previously a simple Map was use when loading Shadows from ShadowProviders.
This means that if an Android class had multiple default shadows, a
ShadowPicker was required to select between them.

The main drawable of a simple Map is not possible to select between shadows
for simple properties like SDK version. This was the case with
ShadowCameraMetaDataNative, which previously required a ShadowPicker to
select which shadow to use based on API level. By using a MultiMap, ShadowMap
itself can use its filter for SDK levels.

It also enables some potential alternatives to ShadowPicker that can have a greatly
reduced amount of boilerplate and increased flexibility.

PiperOrigin-RevId: 450719331