Use API level int for equality and hashing
The equals and hashing methods were previous using `getSdk()` which calls `getContext()` which is not valid for the lifetime of the framework method object (it is used for example in a hashmap by the junit test runner super class to index test descriptions). When the context was destroyed it would start causing the equals and hash methods to start throwing, this can happen for example if there are hash collisions in the framework methods when accessing a hash map keyed using the methods. With a sufficiently large number of tests this could be expected to happen occassionaly.

Instead assume that SDKs are unique per API level, which they always are in practice, and use the API level which is a simple primitive.

PiperOrigin-RevId: 453305228