Cache the helper runners per sandbox
Don't create a new helper runner for every test, just create one per sandbox. In Robolectric the test runner is pretty lightweight but subclasses may wish to store state or perform non-trivial work in helpers, avoid creating new instances where it is unnecessary.

Also factors out the exception handling, by default the runner baseclass can throw an exception on construction so every subclass had to manually handle this when creating the helper runner, instead allow the protected method to just raise the exception and handle it when calling the method.

PiperOrigin-RevId: 453325195