Add an option to use real scrolling logic in ShadowView
This is part of an ongoing effort to
greatly improve the fidelity of graphics-related classes in Robolectric.

Add a system property to disable the fake View scroll methods and call the real code instead.

Longer-term instead of a System property, a new graphics mode will be
used for the check.

Unfortunately too many tests likely depend on the broken scrolling logic for
it to be enabled by default.

PiperOrigin-RevId: 453770196