Return empty lists by default in ShadowAccessibilityManager.get*ServiceList methods
This is consistent with Android, does not return null values for these methods.

This change also has the side effect of enabling Material Snackbar animations.
Material snackbar animations are disabled if the result of
AccessibilityManager.getEnabledAccessibilityServiceList is null. However, these
animations can also be skipped using UiAutomation.setAnimationScale(0f).
and setting a null value for
ShadowAccessibilityManager.setEnabledAccessibilityServiceList is still
supported.

PiperOrigin-RevId: 457778190