Fix ImplementsValidator false positives when looseSignatures is set
Previously, ImplementsValidator would incorrectly label certain methods
as requiring '@Implementation' annotations. This happened when looseSignatures
was set to true, and the method had the same name and the same number of
arguments as one that matches an SDK method.

Fix ImplementsValidator to require that the shadow method use Object as
the param type when looseSignatures is set to true.

PiperOrigin-RevId: 461630677