Implements ShadowUsbRequest and ShadowUsbDeviceConnection
UsbRequest and UsbDeviceConnection allow for communicating with a remote UsbDevice. To play the role of the remote device, we've exposed the methods ShadowUsbDeviceConnection#writeIncomingData and ShadowUsbDeviceConnection#readOutgoingData.

ShadowUsbDeviceConnection#writeIncomingData allows a tester to pass bytes in, that can then be read by an initialized UsbRequest#queue(ByteBuffer).

ShadowUsbDeviceConnection#readOutgoingData allows a tester to confirm bytes that have been written by UsbDeviceConnection#bulkTransfer.

PiperOrigin-RevId: 461690799