Respect permissions granted by `ShadowContextWrapper.grantPermissions()` when sending broadcasts to receivers with permissions.
Broadcast receivers can be registered with a permission that the sender of the broadcast must hold in order for the broadcast to be received. In Robolectric, those permissions are enforced, but the only check that is made is against `PackageManager.checkPermission()`. This only uses data filled in from the application manifest. Unit tests might not have an accurate and complete manifest. For most purposes, this can be worked around by using `ShadowContextWrapper.grantPermissions()` to dynamically grant the needed permission. However, since Robolectric's broadcast receiver mechanism does not currently check the list maintaned by `ShadowContextWrapper`, this does not currently work for broadcast receivers with permissions.

To fix this issue, in addition to checking the manifest-based permissions from `PackageManager`, also do a direct `checkPermission` call on the application context, which will check the dynamic permissions list. The broadcast will go through if either permission check succeeds.

PiperOrigin-RevId: 463020128