Disable EdgeEffect for Android S and above
In Android S and above, ScrollViews will intercept/swallow all touch events
while edge effects are still running. However, it is non-trivial in
Robolectric to complete edge effects in Android S and above. This is because
edge effect anaimations depend on a free-running clock and draw traversals
being continuously performed. So for Robolectric to complete edge effects
that have started, it will need to detect any edge effects that have started,
bump the virtual time, and rerun draw traversals.

Because edge effects are not critical for unit testing, it is simpler to
disable them.

This fixes an issue with swiping up in a ScrollView in Espresso and then trying
to click a child view of the ScrollView.

PiperOrigin-RevId: 463097071