Avoid calling reflector as part of object finalization
When reflector(..) is called, it will generate a new reference to the object
being invoked. If this happens in a finalizer, it will delay GC for that object
in a recursive manner, and will result in the finalizer being invoked
repeatedly.

Use `@ReflectorObject` to cache the reflector object to avoid creating new
object references.

This was discovered by telemetry to the 'reflector' method that logs an event
if the reflector object is created in a thread named 'finalizer'.

PiperOrigin-RevId: 463113555