Work around race condition in CursorWindow.dispose that can lead to segfaults
On the JVM, it is possible to have two finalizer threads running concurrently.
This can happen if System.runFinalization is called, which starts a secondary
finalizer thread. When this happens, it is possible for CursorWindow's
'dispose' method to be called concurrently, leading to a double-free of native
resources which can cause a segfault. Add protection against this situation by
using a simple AtomicBoolean.

PiperOrigin-RevId: 465357165