Avoid calling the real DisplayManagerGlobal constructor in RNG
The real DisplayManagerGlobal constructor references the ColorSpace named
constants, which require native calls to instantiate. Because
DisplayManagerGlobal is initialized when an Application is created, this will
cause native graphics libraries to always be loaded when an Application is
created. Instead override the constructor to avoid referencing the ColorSpace
named constants.

This makes tests that create application objects but do not use graphics around
~0.75s faster.

PiperOrigin-RevId: 469859159