Store CallAudioState in FakeTelecomServer, instead of storing across various Shadow classes.
Up until now, manipulating CallAudioState in Robolectric unit tests has not been straightforward. The state of the current call audio is stored across ShadowInCallService, ShadowInCallAdapter, and ShadowPhone.

We now have a new solution in which we bind an InCallAdapter implementation to the InCallService in our unit test. In real Telecom code, this Adapter is where the Phone application sends IPC to the underlying Telecom server implementation, at the OS level.

Now, all invocations of InCallService#setMuted and InCallService#setAudioRoute are received by real Telecom objects, and are eventually intercepted by the InCallAdapter implementation. This InCallAdapter now contains all Call Audio state.

I had to use some reflection to check if we are using the new implementation of an InCallAdapter, so it would not break the previously used implementation.

PiperOrigin-RevId: 470031909