Set the default ICU Locale when the native runtime is loaded
This is required to make Android SQLite's 'COLLATE UNICODE' extension work. For
'COLLATE UNICODE', the Android SQLite extension calls ucol_open with NULL as
the locale, which causes the default locale rules to be used. Because
previously the default ICU locale was never set, it wouldn't use rules
consistent with real Android. In real Android the native ICU locale is kept in
sync with the Java Locale. For the purposes of RNR setting the default locale
upfront should be sufficient, as hooking into Locale.setDefault would require
attaching a Java agent.

https://cs.android.com/android/platform/superproject/+/master:external/sqlite/android/sqlite3_android.cpp

PiperOrigin-RevId: 471819495