Avoid inlining StatusBarManager.DEFAULT_SETUP_DISABLE{,2}_FLAGS
Previously ShadowStatusBarManager referenced the StatusBarManager default
disable constants in a way that caused them to be inlined. In Android T the
value of DEFAULT_SETUP_DISABLE2_FLAGS changed, and this value was getting
inlined in the ShadowStatusBarManager class file. Instead use reflector to
reference the runtime value to ensure it has the correct value.

PiperOrigin-RevId: 473089301