Parse `idx` and `offset` correctly from sparse resource entries
The data structure is:

```
union ResTable_sparseTypeEntry {
    // Holds the raw uint32_t encoded value. Do not read this.
    uint32_t entry;
    struct {
        // The index of the entry.
        uint16_t idx;

        // The offset from ResTable_type::entriesStart, divided by 4.
        uint16_t offset;
    };
};
```

This has been incorrectly transcribed as if it is an outer struct with an inner union (i.e. an int and a short which is either index or offset), when in practice it should be a union of an int and a struct of two shorts, where the int is essentially unused.

In sparse resource table each entry is 4 bytes, the first 2 bytes are the index and the second 2 bytes are the offset.

Correctly read the index and offset and don't bother representing the entry.

PiperOrigin-RevId: 474220169