Restrict the use of getDeclaredFields0 in ReflectionHelpers
Currently, any call to ReflectionHelpers field methods can result in the use of
getDeclaredFields0, even when the field is accessible. This prints a scary
log message, which is most noticeable in a gradle environment.

The use of getDeclaredFields0 should only be necessary when setting final
fields, and this will be phased out eventually.

Update ReflectionHelpers to avoid using getDeclaredFields0 by default.

PiperOrigin-RevId: 475475246