Make GC of native SQLiteConnection objects thread-safe
This was the root cause of the segfaults when turning on native SQLite by
default. A finalizer thread was running concurrently with the legacy 'close'
method due, possibly due to the threading model of legacy SQLite, causing a
segfault.

Add synchronization logic to ShadowSQLiteConnection.dispose, and move up the
CursorWindow dispose synchronization to ShadowCursorWindow so it can be reused
with legacy CursorWindow.

PiperOrigin-RevId: 476221241