Create check to see if constructor can be instrumented
If constructor bytecode is unsupported, skip instrumenting it.

This occurs when trying to instrument the GMSCore ApiException class from
play-services-basement, which is desugared and proguarded.
That constructor begins with aload_1 and contains additional instructions
after RETURN, which is not supported at the moment.

A quick heuristic to check if a constructor can be instrumented is checking
for an aload_0 as the first instruction. This always happens when Android
framework code is compiled using javac.

Fixes #7269

PiperOrigin-RevId: 477284726