Remove overly broad catch(Exception) from JarInstrumentor
There was a try/catch statement that caught all Exceptions when instrumenting classes and does not throw an error. This makes debugging more difficult and also hides away classes that do fail to instrument. We have isolated the known crashes to be a NegativeArraySizeException and are now explicitly catching NegativeArraySizeException so that any other Exceptions that we do not know of (for example, when writing new code), can be thrown.

PiperOrigin-RevId: 480149705