Fix loading Robolectric Native Runtime with concurrerent threads
Previously, the logic to load the Robolectric Native Runtime used
a simple AtomicBoolean to keep track of whether process to load
the shared object file had been started. This meant that if this logic
was initiated by two threads, the loser of the race would likely attempt
to call native methods that hadn't been bound yet.

Update the lazy-loading logic to be triggered when 'entry point' shadow
methods are called. These are typically constructors and finalizer-related
methods, though they may vary for each different native-backed object.

Also update the RNR loading logic to be synchronized. It is safe to do this now
as the lazy loading logic is triggered from shadows, so a static initializer
deadlock won't result.

PiperOrigin-RevId: 480523320