Remove ShadowDrawable corrupt source APIs
These APIs are unused and misguided. Because Drawable.createFromStream is
incorrectly shadowed to always returned a Bitmap object, regardless of the
contents of the stream, this API was the only way to have a null value
returned.

In a forthcoming change, the Drawable.createFromStream shadow will be deleted,
so it will be possible to test corrupt streams using the real
Drawable.createFromStream code.

PiperOrigin-RevId: 481696703