Remove ShadowDrawable.createFromPath
Previously in Robolectric, Drawable.createFromPath always returned
a Drawable object, even if the path was totally invalid. This is
not realistic as in real Android, as Drawable.createfFromPath will return
null if the path is invalid or does not contain real Bitmap data.

Remove ShadowDrawable.createfFromPath to defer to the real code instead.

PiperOrigin-RevId: 481999530