Make ShadowCanvas public APIs extendable
ShadowCanvas has a lot of public interaction/verify APIs that are necessary
because in Robolectric Canvas is a no-op object. It is not currently possible
to make assertions on the pixel results of Canvas operations.

Longer-term there will be an alternative to ShadowCanvas that is backed by real
Android native graphics libraries, such as Skia, Minikin, and the various image
libraries.

Update ShadowCanvas to be the base Canvas class that only contains the public
API. This is similar to ShadowLooper. This makes it possible to extend the
ShadowCanvas API without extending the shadow itself.

PiperOrigin-RevId: 483426552