Treat `Pattern` as the same if the pattern and flags are the same
The rule puts all expected log messages into a hash map to de-deuplicate but `Pattern` does not implement an equals/hash method that check equality of the regex and flags used to compile the object, it just uses default object equality. Instead test `Pattern` as identical if they use the same regex/flags.

PiperOrigin-RevId: 485348987