Change where addCallToRoboInit() inserts call to $$robo$init
addCallToRoboInit() is used to add a $$robo$init call to constructors that belong to classes that are instrumented by Jacoco (the $$robo$init method instantiates the class' shadow object). The addCallToRoboInit() method currently inserts the call to $$robo$init before the return statement. However, it is not guaranteed that a constructor will have a return statement. This change will make it so that the addCallToRoboInit() method adds the call to $$robo$init after the call to super() or this().

PiperOrigin-RevId: 486690877