Update ShadowAlarmManager to support missing APIs, and add new behavior to schedule and run alarms based on the main Looper time, guarded by ShadowAlarmManager.setAutoSchedule(). This could potentially become the default in the future.
If auto-scheduling is enabled, when an alarm is set and the main looper time is advanced to the alarm trigger time, the alarm will be fired (PendingIntent will be sent or OnAlarmListener will be invoked) on the proper thread. This allows for better black box unit testing of alarms without requiring encapsulation leakiness in prod code.

Also adds a new fireAlarm() method which manually fires an alarm - this can be used to emulate reordered or delayed alarms (as may happen on a real device).

PiperOrigin-RevId: 487589674