Avoid per-sdk logic in WindowSessionDelegate
The InsetsVisibility class is going away in U, so the per-sdk approach will not
be viable as the S_V2 and T delegates contains InsetsVisibility args.

Update to use Proxy.newProxyInstance directly which will provide
more flexibility in intercepting method calls based on the name of
the method and not the exact argument types.

PiperOrigin-RevId: 491648117