Introduce a set of graphics shadows backed by native code
Previously, Android graphics classes in Robolectric were mostly no-ops. There
were also some low-fidelity fakes, such as Bitmap and BitmapFactory, backed by
Java code or AWT.

This is the first of a series of commits that will enable a native graphics
mode for Robolectric. This will be available on Linux and Mac for SDK versions
O (26)through T (33). Add a large set of native-backed shadows and binding
classes that will invoke the Android native graphics libraries. This mode can
be enabled using the `@GraphicsMode(GraphicsMode.Mode.NATIVE)` annotation.

Currently the native graphics libraries and related artifacts are built on a
custom Android S (v12) branch in AOSP. There will be some forthcoming commits
that will enable building these into the nativeruntime jar. Once this is
possible, a large number of tests will be added to verify the behavior.

PiperOrigin-RevId: 494946511